+++ dbt_platform_helper/exceptions.py
@@ -18,3 +18,23 @@ class IncompatibleMinorVersion(ValidationException):
         super().__init__()
         self.app_version = app_version
         self.check_version = check_version
+
+
+class CopilotCodebaseNotFoundError(Exception):
+    pass
+
+
+class NotInCodeBaseRepositoryError(Exception):
+    pass
+
+
+class NoCopilotCodebasesFoundError(Exception):
+    pass
+
+
+class ImageNotFoundError(Exception):
+    pass
+
+
+class ApplicationDeploymentNotTriggered(Exception):
+    pass
